requirements.txt
Flask
Flask_Cors==3.0.10
Flask_Session==0.4.0
lxml
redis
redis==4.5.4
Requests
SQLAlchemy
Werkzeug

Pillow
PyYAML

torch
torchvision
blinker
